Enable update of the tab name of Drive views in the sidebar
